+++ gob_import.py
@@ -26,43 +26,15 @@ from struct import unpack
 import string
 from bpy.types import Operator
 from bpy.props import EnumProperty
-from bpy.app.translations import pgettext_iface as iface_
-
 from . import geometry, output, paths, utils, nodes
 
 
 run_background_update = False
-icons = None
-cached_last_edition_time = time.perf_counter()
-last_cache = 0
-preview_collections = {}
 gob_import_cache = []
-
-
-def draw_goz_buttons(self, context):
-    global run_background_update, icons
-    if context.region.alignment == 'RIGHT':
-        layout = self.layout
-        row = layout.row(align=True)
-
-        icons = preview_collections["main"]
-
-        if utils.prefs().show_button_text:
-            row.operator(operator="scene.gob_export_button", text="Export", emboss=True, icon_value=icons["GOZ_SEND"].icon_id)
-            if run_background_update:
-                row.operator(operator="scene.gob_import", text=iface_("Import", None), emboss=True, depress=True, icon_value=icons["GOZ_SYNC_ENABLED"].icon_id).action = 'AUTO'
-            else:
-                row.operator(operator="scene.gob_import", text=iface_("Import", None), emboss=True, depress=False, icon_value=icons["GOZ_SYNC_DISABLED"].icon_id).action = 'AUTO'
-            row.operator(operator="scene.gob_import", text="Manual", emboss=True, depress=False, icon='IMPORT').action = 'MANUAL'
-        else:
-            row.operator(operator="scene.gob_export_button", text="", emboss=True, icon_value=icons["GOZ_SEND"].icon_id)
-            if run_background_update:
-                row.operator(operator="scene.gob_import", text="", emboss=True, depress=True, icon_value=icons["GOZ_SYNC_ENABLED"].icon_id).action = 'AUTO'
-            else:
-                row.operator(operator="scene.gob_import", text="", emboss=True, depress=False, icon_value=icons["GOZ_SYNC_DISABLED"].icon_id).action = 'AUTO'
-            row.operator(operator="scene.gob_import", text="", emboss=True, depress=False, icon='IMPORT').action = 'MANUAL'
-
+cached_last_edition_time = time.perf_counter()
 start_time = None
+
+
 class GoB_OT_import(Operator):
     bl_idname = "scene.gob_import"
     bl_label = "Import from GOZ"
@@ -731,55 +703,6 @@ class GoB_OT_import(Operator):
             return{'FINISHED'}
 
 
-
-       
-class GoB_OT_export_button(Operator):
-    bl_idname = "scene.gob_export_button"
-    bl_label = "Export to ZBrush"
-    bl_description = "Export selected Objects to ZBrush\n"\
-                        "LeftMouse: as Subtool\n"\
-                        "SHIFT/CTRL/ALT + LeftMouse: as Tool"
-    bl_options = {'INTERNAL'}
-    
-    @classmethod
-    def poll(cls, context):              
-        return geometry.export_poll(cls, context)
-        #return bpy.context.selected_objects
-
-    def invoke(self, context, event):
-        as_tool = event.shift or event.ctrl or event.alt
-        bpy.ops.scene.gob_export(as_tool=as_tool)
-        return {'FINISHED'}
-
-
-
-
-
-class GOB_OT_Popup(Operator):
-    bl_label = "GoB: Zbrush Path not found!"
-    bl_description ="look for zbrush in specified path or in default installation"\
-                    "directories and if its not found prompt the user to input the path manually"
-    bl_idname = "gob.search_zbrush"
-          
-    def draw(self, context):        
-        self.layout.label(text='Please set your ZBrush path', icon='COLORSET_01_VEC')
-        self.layout.label(text='        in the Add-ons Preferences')
-
-    def open_addon_prefs(self, context):
-        context.window_manager.addon_support = {'OFFICIAL', 'COMMUNITY', 'TESTING'}
-        bpy.ops.preferences.addon_show(module=__package__)
-
-    def invoke(self, context, event):       
-        wm = context.window_manager
-        font_size_correction = bpy.context.preferences.ui_styles[0].widget_label.points / 10
-        return wm.invoke_props_dialog(self, width = int(200 * font_size_correction))
-
-    def execute(self, context):
-        self.open_addon_prefs(context)
-        return {'FINISHED'}
-
-
-
 def run_import_periodically():
     # print("Runing timers update check")
     global cached_last_edition_time, run_background_update
@@ -814,7 +737,6 @@ def run_import_periodically():
         bpy.app.timers.unregister(run_import_periodically)
         
     return utils.prefs().import_timer
-
 
 
 def run_import_manually():
